+++ src/Resource/Segment.php
@@ -8,15 +8,6 @@ use Collinped\Aimtell\Aimtell;
 class Segment extends BaseResource
 {
     /**
-     * Protect assigned methods from being executed by a resource.
-     *
-     * @var array
-     */
-    protected array $guarded = [
-        'getResultsByDate',
-    ];
-
-    /**
      * Segment constructor.
      * @param Aimtell $aimtell
      * @param string|null $segmentId