+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `aimtell-php` will be documented in this file.
 
+## 1.0.7 - 2020-12-28
+
+- Added Welcome Notifications
+- Updated documentation links to Aimtell Docs
+- Merged bugfixes for SDK
+
 ## 1.0.0 - 2020-12-16
 
 - Initial release